Added function in local and tidy up the ui and config the data in DetailBookActivity.kt

+++ app/src/main/java/com/b21/finalproject/smartlibraryapp/ui/home/ui/home/HomeAdapter.kt
@@ -67,7 +67,7 @@ class HomeAdapter : RecyclerView.Adapter<HomeAdapter.ViewHolder>() {
 
                 itemView.setOnClickListener {
                     val intent = Intent(itemView.context, DetailBookActivity::class.java)
-                    intent.putExtra(DetailBookActivity.BOOK_ID, bookEntity.bookId)
+                    intent.putExtra(DetailBookActivity.BOOK_ID, bookEntity.bookId.toInt())
                     itemView.context.startActivity(intent)
                 }
             }
